Fixed accessiblity for Cart & menu items, Responsivness Fix for Product Page
