+++ src/index.html
@@ -10,6 +10,7 @@
         <meta name="description" content="Sabka Bazar is a mini cart assignment">
     </head>
     <body>
+        <noscript>You need to enable JavaScript to run this app.</noscript>
         <div id="root"></div>
     </body>
 </html>
