+++ src/styles/_global.scss
@@ -2,6 +2,8 @@
 
 $bgColor: #e91e63;
 $greyColor: #ddd;
+$headerHeightDesktop: 100px;
+$headerHeightMobile: 70px;
 
 * {
     padding: 0;
@@ -31,14 +33,12 @@ p {
 
 .container.product_area_container {
     width: 100%;
-    padding-right: 15px;
-    padding-left: 15px;
+    padding: 0 10% 0 0%;
     margin-right: auto;
     margin-left: 250px;
-    /*max-width: 1140px;*/
     margin-top: 7rem;
     padding-bottom: 15px;
-    max-width: calc(100% - 250px);
+    min-height: 100vh;
 }
 
 .container {
@@ -118,7 +118,9 @@ p {
     padding-top: 1rem;
 }
 
-
+.home-sec{
+    min-height: 90vh;
+}
 
 .slider_container {
     width: 100%;
@@ -130,14 +132,14 @@ p {
     max-width: 1390px;
     height: fit-content;
     box-shadow: 0 13px 25px -23px #a5a5a5ad;
-    margin-top: 8rem;
+    margin-top: $headerHeightDesktop + 15px;
 }
 
 footer {
-    position: relative;
+    /*position: relative;*/
     background-color: $greyColor;
     padding: 1rem 0;
-    z-index: 100;
+    /*z-index: 98;*/
 
     p {
         margin: 0;
@@ -212,91 +214,6 @@ footer {
 
 .hamburger-button{
     display: none;
-}
-
-.top_header {
-    padding: 0.5rem 0;
-    position: fixed;
-    width: 100%;
-    top: 0px;
-    background: #fff;
-    z-index: 100;
-
-    ul {
-        padding: 0;
-        margin: 0 -15px;
-        list-style: none;
-        display: flex;
-        justify-content: flex-end;
-
-        li {
-            margin: 0 0 0 1rem;
-
-            a {
-                color: #282828;
-                text-decoration: none;
-            }
-        }
-    }
-}
-
-header {
-    padding: 0.5rem 0;
-    box-shadow: 0 15px 15px -21px #737373;
-    position: fixed;
-    top: 1.6rem;
-    background: #fff;
-    width: 100%;
-    z-index: 100;
-    .menu_list {
-        list-style: none;
-        padding: 0;
-        margin: 0;
-        display: flex;
-        justify-content: flex-start;
-        margin-top: 0.4rem;
-
-        li {
-            a {
-                text-decoration: none;
-                padding: 1rem;
-                display: block;
-                color: #888;
-
-                &:hover {
-                    color: #111;
-                }
-            }
-        }
-    }
-
-    .nav_logo {
-        img {
-            width: 120px;
-        }
-    }
-
-    .nav_cart {
-        text-align: right;
-        cursor: pointer;
-
-        p {
-            background-color: $greyColor;
-            display: inline-block;
-            padding: 1rem 1.5rem;
-
-            img {
-                width: 20px;
-                position: relative;
-                top: 5px;
-                margin-right: 4px;
-            }
-        }
-    }
-}
-
-.top-container{
-    margin-top: 7rem;
 }
 
 .close-btn-container.hidden{
@@ -393,9 +310,11 @@ textarea {
   }
 
 /* ========= INPUTS ======= */
-  
+.products-filter-mobile{
+    display: none;
+}
 
-@media (min-width: 768px) and (max-width: 1023px) {
+@media (min-width: 768px){
 
     .banner_card {
         .product_content {
@@ -409,20 +328,6 @@ textarea {
 }
 
 @media (max-width: 767px) {
-
-    .hamburger-button{
-        background: transparent;
-        padding: 0.5rem;
-        margin: 0.5rem;
-        display: block;
-        float: left;
-        width: 40px;
-        height: 38px;
-        border: 1px solid #c1c1c1;
-    }
-    .hamburger-button:hover{
-        cursor: pointer;
-    }
     
     .container {
         .row {
@@ -441,10 +346,36 @@ textarea {
         }
     }
     
+    .slider_container{
+        padding: 0;
+    }
 
     .banner_card {
+        .col.col-md-4{
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            flex: 1;
+        }
+        .col.col-md-8{
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            flex: 1;
+        }
         .product_content {
             margin-top: 0;
+            h2{
+                font-size: 18px;
+            }
+            p{
+                font-size: 11px;
+                padding: 15px 3px;
+            }
+            .btn-primary{
+                font-size: 10px;
+                font-weight: bold;;
+            }
         }
 
         .product_img {
@@ -458,76 +389,29 @@ textarea {
         text-align: center;
     }
 
-    header {
-        top: 0px;
-        .container {
-            .row {
-                .col-md-3 {
-                    -ms-flex: 0 0 50%;
-                    flex: 0 0 50%;
-                    max-width: 50%;
-                }
-
-                .col-md-6 {
-                    display: none;
-                }
-            }
-        }
-    }
-    .top_header {
-        display: none;
-    }
-
     .container.product_area_container{
+        padding: 0px;
         max-width: 100%;
-    }
-    
-    nav.main-navigation.open{
-        display: block;
-        position: fixed;
-        width: 250px;
-        height: 100%;
-        background-color: #eaeaea;
-        transition: 0.5s;
-        overflow: hidden;
-        margin-left: 0rem;
-        top: 4.8rem;
-        z-index: 93;
-        .close-btn-container{
-            display: block;
-            padding: 1rem 0.5rem;
-            button.\.close{
-                float: right;
-                padding: 0.5rem;
-                border: 0px;
-                background: #eaeaea;
-                font-weight: bolder;
-                cursor: pointer;
-            }   
-        }
-        ul.menu_list {
-            display: block;
-            top: 0;
-            left: 0;
-            width: 100%;
-            li {
-                position: relative;
-                width: 100%;
-                list-style: none;
-                border-bottom: 1px solid #d1d1d1;
-                &:hover {
-                    background-color: #f9f8f8;
-                    cursor: pointer;
-                }
-                
-                &.active{
-                    background-color: #e91e63;
-                    .item-title{
-                        color: #FFF;
-                    }
-                }
-            }
-        }
     }
 }
 
+@media (max-width: 600px) {
+    .container.product_area_container{
+        margin-top: 15px;
+    }
+    .products-filter-mobile {
+        display: block;
+        margin-top: 70px;
+        margin-left: -15px;
+        margin-right: -15px;
+        .products-filter-select-mobile {
+            width: 100%;
+            border: 0px;
+            padding: 10px;
+            height: 45px;
+            background: #e91e63;
+            color: white;
+            border: 0px;
+        }
+    }
+}
