+++ src/styles/_variables.scss
@@ -2,3 +2,6 @@
     --text: #fff;
 }
 $bg: black;
+$headerHeightDesktop: 100px;
+$headerHeightMobile: 70px;
+$fontFamily: "Ropa Sans", sans-serif;
