+++ src/styles/product-card.scss
@@ -4,10 +4,10 @@
     flex-wrap: wrap;
     grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
     grid-auto-rows: minmax(150px, auto);
-    grid-gap: 1em;
+    grid-gap: 2em;
     align-items: stretch;
   }
-  
+
   .product-card {
     display: flex;
     align-items: center;
@@ -17,10 +17,9 @@
     flex-wrap: wrap;
     border-bottom: 2px dashed #eaeaea;
     box-shadow: 3px 0px 5px -3px #f1ebebe3;
-    margin-left: 5px;
-    margin-right: 5px;
-    flex: 1 1 200px;
+    flex: 1 1 270px;
     transition: box-shadow 0.2s ease-in-out;
+    max-width: 270px;
     &:hover{
       box-shadow: 0px 0px 10px 4px #e6e5e5a6;
     }
@@ -39,6 +38,7 @@
     text-overflow: ellipsis;
     overflow: hidden;
     height: 7rem;
+    font-size: 0.916rem;
   }
 
   .product-card .product-title h3{
@@ -73,24 +73,166 @@
     top: 15%;
   }
 
-@media (max-width: 767px) {
-  .product-card .product-image {
-    margin: 0px auto;
-    float: left;
+  .product-card .product-info-mobile{
+    display: none;
+  }
+  
+.product-card.tablet{
+  display: none;
+}
+
+.product-card.mobile{
+  display: none;
+}
+.container.product_area_container{
+  padding: 0px 0px 0px 15px;
+}
+
+@media only screen and (max-width: 769px) {
+  section.product_sec {
+    margin-top: -1.9rem;
+  }
+  .grid{
+    width: 100%;
     display: flex;
-    max-width: 40%;
+    flex-wrap: wrap;
+    grid-template-columns: auto;
+    grid-auto-rows: auto;
+    grid-gap: 2em;
     flex-direction: row;
     align-items: stretch;
+    justify-content: flex-start;
   }
-
-  .product-card .product-short-description {
-    float: right;
+  .sidebar{
+    left: 0px;
+    margin-left: -1rem;
+    max-width: 200px;
+  }
+  .container.product_area_container{
+    margin-left: 180px;
+  }
+  .product-card.desktop{
+    display: none;
+  }
+  
+  .product-card.mobile{
+    display: none;
+  }
+  .product-card.tablet{
     display: flex;
-    flex-direction: row;
-    max-width: 55%;
-    margin-left: 5%;
-    span {
+    justify-content: flex-start;
+    padding: 15px;
+    flex-wrap: wrap;
+    flex: none;
+    max-width: 250px;
+    box-shadow: none;
+    &:hover{
+      box-shadow: 0px 0px 10px 4px #e6e5e5a6;
+    }
+    .product-title h3{
+      font-size: 1rem;
+      padding: 15px 0px 15px 0px;
+    }
+    .product-info-container{
+      display: flex;
+      flex-direction: row;
+      .product-image-container{
+        justify-content: stretch;
+        flex-grow: 1;
+        flex-basis: 0;
+        display: flex;
+        align-items: center;
+        img{
+          width: 100%;
+        }
+      }
+      .product-short-description{
+        flex-grow: 1;
+        flex-basis: 0;
         overflow: hidden;
+        font-size: 0.87rem;
+      }
+    }
+    .buy-button-container{
+      width: 100%;
+      margin-top:0.7rem;
+      button{
+        font-size: 0.8rem;
+        width: 100%;
+        padding: 0.5rem;
+      }
+    }
+  } 
+}
+
+@media  only screen and (max-width: 600px) {
+  section.product_sec {
+    margin-top: 0;
+  }
+  .grid{
+    display: block;
+    width: 100%;
+  }
+  .product-card.tablet{
+    display: none;
+  }
+  .product-card.desktop{
+    display: none;
+  }
+  .container.product_area_container{
+    margin-left: 0; 
+  }
+  .product-card.mobile{
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    height: auto;
+    padding: 15px;
+    flex-wrap: wrap;
+    border-bottom: 2px dashed #eaeaea;
+    flex: none;
+    max-width: 100%;
+    box-shadow: none;
+    &:hover{
+      box-shadow: 0px 0px 10px 4px #e6e5e5a6;
+    }
+    .product-title h3{
+      padding: 15px 0px 15px 0px;
+      font-size: 0.95rem;
+    }
+    .product-info-container{
+      display: flex;
+      flex-direction: row;
+      gap: 0.3rem;
+      .product-image-container{
+        width: 100%;
+        display: flex;
+        align-items: center;
+        justify-content: stretch;
+        img{
+          width: 100%;
+        }
+      }
+      .product-desc-container{
+        display: flex;
+        flex-direction: column;
+        align-items: stretch;
+        justify-content: center;
+        gap: 0.5rem;
+        .product-short-description{
+          display: felx;
+          flex-wrap: wrap;
+          max-width: 100%;
+          margin: none;
+          height: auto;
+          font-size: 0.75rem;
+        }
+        .buy-button-container button{
+          width: 100%;
+          font-size: 0.8rem;
+          padding: 0.5rem;
+        }
+      }
     }
   }
 }
