+++ src/styles/slick-theme.scss
@@ -164,3 +164,17 @@ $slick-opacity-not-active: 0.25 !default;
         }
     }
 }
+
+/* Extra small devices (phones, 600px and down) */
+@media only screen and (max-width: 600px) {
+    .slick-prev {
+        left: -5px;
+        z-index: 1;
+    }
+    .slick-next {
+        right: 1.1rem;
+    }
+    .slick-prev, .slick-next{
+        display: none;
+    }
+}
